% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{DE_MAST_RE_seurat}
\alias{DE_MAST_RE_seurat}
\title{MAST differential expression test with random effect for Seurat object}
\usage{
DE_MAST_RE_seurat(
  object,
  random_effect.vars,
  ident.1 = NULL,
  ident.2 = NULL,
  cells.1 = NULL,
  cells.2 = NULL,
  group.by = NULL,
  logfc.threshold = 0.25,
  base = exp(1),
  assay = NULL,
  slot = "data",
  features = NULL,
  min.pct = 0.1,
  max.cells.per.ident = NULL,
  random.seed = 1,
  latent.vars = NULL,
  n_cores = NULL,
  verbose = TRUE,
  p.adjust.method = "fdr",
  ...
)
}
\arguments{
\item{object}{Seurat >=3 object, with log-normalized data in the `data` slot}

\item{random_effect.vars}{Character vector of variables to add as random intersects in MAST model i.e. ~ ...  + (1|random_effect.var1) + (1|random_effect.var2)}

\item{ident.1}{Identity class to define markers for}

\item{ident.2}{A second identity class for comparison. Leave as NULL to compare with all other cells}

\item{cells.1}{Vector of cell names belonging to group 1. Alternative way to specify ident.1}

\item{cells.2}{Vector of cell names belonging to group 2. Alternative way to specify ident.2}

\item{group.by}{Regroup cells into a different identity class prior to performing differential expression}

\item{logfc.threshold}{Only return results with a DE exceeding threshold}

\item{base}{base for log when computing mean and for output, default exp(1). NB: Seurat has changed to log2 in V4.}

\item{assay}{Assay to pull data from; defaults to default assay}

\item{slot}{Slot to pull data from; defaults to "data" as MAST expects log-normalized data}

\item{features}{Genes to test. Default is to use all genes}

\item{min.pct}{Only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expressed. Default is 0.1}

\item{max.cells.per.ident}{Downsample each identity class to a max number. Default is no downsampling. Not activated by default (set to Inf)}

\item{random.seed}{Random seed to use for down sampling}

\item{latent.vars}{Variables to test}

\item{n_cores}{How many cores to use (temporarily sets options(mc.cores=n_cores))}

\item{verbose}{whether to print stdout from MAST functions}

\item{p.adjust.method}{passed to p.adjust. Note that n is all the genes in the object}

\item{\dots}{Additional parameters (other than formula, sca, method, ebayes, strictConvergence) to pass to MAST::zlm}
}
\value{
data.frame with column names p_val, avg_log[base]FC, pct.1, pct.2, p_val_adj (identical format to Seurat::FindMarkers)
}
\description{
Uses Seurat::FindMarkers MAST test hurdle model with added random effects variables
}
\references{
Zimmerman, K.D., Espeland, M.A. & Langefeld, C.D.
. A practical solution to pseudoreplication bias in single-cell studies.
. Nat Commun 12, 738 (2021). https://doi.org/10.1038/s41467-021-21038-1

. McDavid A, Finak G, Yajima M (2020). MAST: Model-based Analysis of Single Cell Transcriptomics. R package version 1.16.0, https://github.com/RGLab/MAST/.

. Stuart and Butler et al. Comprehensive Integration of Single-Cell Data. Cell (2019)

. Seurat 3 source at https://github.com/satijalab/seurat/blob/master/R/differential_expression.R
}
