% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{.countsperclus}
\alias{.countsperclus}
\title{Compute counts per cluster}
\usage{
.countsperclus(object, group, assay = NULL, slot = NULL, min.cell = 100)
}
\arguments{
\item{group}{define groups which you want to split; default is srt ident}

\item{srt}{seurat object}

\item{compare}{vector of groups you want to compare #add in}
}
\description{
subroutine
}
\details{
.. content for \description{
goal of this function is to find genes that are specifically expressed in each cluster, it is also
significantly faster than the FindMarkers function from Seurat
} (no empty lines) ..

.. content for \details{
pseudobulk calculation inspired by: https://jef.works/blog/2020/04/06/quickly-creating-pseudobulks/
weighted log odds from: https://github.com/juliasilge/tidylo
}
}
\author{
dylanmr
}
