% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{.dub_cutoff}
\alias{.dub_cutoff}
\title{A subroutine of prep_intrahash_doub to compute number of doublet neighbours}
\usage{
.dub_cutoff(x)
}
\arguments{
\item{x}{an SingleCellExperiment object}
}
\value{
a logical vector of prediction scores
}
\description{
A subroutine of prep_intrahash_doub to compute number of doublet neighbours
}
