% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_dimensions.R
\name{estimate_dimensions}
\alias{estimate_dimensions}
\title{Estimate dataset dimensionality}
\usage{
estimate_dimensions(x, max_dim = 75, seed.use = 12345)
}
\arguments{
\item{x}{: Seurat or SingleCellExperiment object}

\item{max_dim}{: max number of dimensions to try}

\item{seed.use}{: random seed}
}
\value{
integer number of dimensions, <= max_dim
}
\description{
Estimate dataset dimensionality
}
\examples{
##estimate_dimensions(seurat_obj)
}
