% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{FoldChange.default}
\alias{FoldChange.default}
\title{Compute average log fold changes}
\usage{
\method{FoldChange}{default}(data, cells.1, cells.2, mean.fxn, fc.name, features = NULL, ...)
}
\arguments{
\item{data}{feature * sample matrix}

\item{cells.1}{Vector of cell names belonging to group 1}

\item{cells.2}{Vector of cell names belonging to group 2}

\item{features}{Features to calculate fold change for.
If NULL, use all features}
}
\description{
Compute average log fold changes
}
\references{
Copied from Seurat 3 source at https://github.com/satijalab/seurat/blob/master/R/differential_expression.R
}
\concept{differential_expression}
