% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset.R
\name{geneset_embed}
\alias{geneset_embed}
\title{Generate activity scores (embeddings) for a feature set in a sample matrix}
\usage{
geneset_embed(mat_datExpr, vec_geneWeights, min_feats_present = 5)
}
\arguments{
\item{mat_datExpr}{a m feature * n sample matrix}

\item{vec_geneWeights}{a numeric vector of weights with feature names. Weights should be pre-normalized}

\item{min_feats_present}{a minimum number of features shared between mat_datExpr and vec_geneWeights}
}
\value{
returns a n vector of gene weights, scaled to 1 standard deviation, but not centred.
}
\description{
Generate activity scores (embeddings) for a feature set in a sample matrix
}
