% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset.R
\name{geneset_embed_list_seurat}
\alias{geneset_embed_list_seurat}
\title{Generate activity scores (embeddings) for a list of feature sets and add to a Seurat object}
\usage{
geneset_embed_list_seurat(
  seurat_obj,
  list_vec_geneWeights,
  slot = "scale.data",
  assay = "SCT",
  min_feats_present = 5,
  n_cores_max = Inf
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{list_vec_geneWeights}{a list of numeric vectors of weights with feature names. Weights should be pre-normalized}

\item{slot}{Seurat object slot to use}

\item{assay}{Seurat object assay to use}

\item{min_feats_present}{a minimum number of features shared between mat_datExpr and vec_geneWeights}

\item{n_cores_max}{max number of cores to use in multicore processing. To use sequential processing, set to 1}
}
\value{
a Seurat object with feature embeddings added to metadata
}
\description{
A variant of geneset_embed that works with a list of feature sets and takes and returns a Seurat object
}
