% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_network}
\alias{plot_network}
\title{make a network plot of a set of co-expressed features}
\usage{
plot_network(
  mat_datExpr,
  vec_geneImportance,
  vec_genes_highlight = c(),
  n_max_genes = 50,
  igraph_algorithm = "drl",
  fontface_labels = "bold.italic",
  color_edge = "grey70",
  edge_thickness = 1
)
}
\arguments{
\item{mat_datExpr}{gene * sample expression matrix}

\item{vec_geneImportance}{vector of weights with corresponding features as names}

\item{vec_genes_highlight}{optional vector of genes to highlight}

\item{n_max_genes}{max number of genes to include, in order of weight}

\item{igraph_algorithm}{passed as the algorithm argument to ggraph()}

\item{fontface_labels}{one of 'normal', 'bold', 'italic', 'bold italic'}

\item{color_edge}{color for network edges, character}

\item{edge_thickness}{controls edge thickness, which varies by co-expression}
}
\value{

}
\description{
make a network plot of a set of co-expressed features
}
