% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prep_HTO_q_area_plot}
\alias{prep_HTO_q_area_plot}
\title{Make an area plot of proportions of cells assigned to singlets, doublets, and negatives
as a function of the positive.quantile provided to Seurat::HTODemux}
\usage{
prep_HTO_q_area_plot(
  seurat_obj,
  vec_range_quantile = seq(0.8, 0.99, 0.01),
  n_cores_max = Inf
)
}
\arguments{
\item{seurat_obj}{A seurat object}

\item{vec_range_quantile}{a numeric vector of quantile values (0-1)}

\item{n_cores_max}{max number of cores to use in multicore processing}
}
\value{
a ggplot2 object
}
\description{
Make an area plot of proportions of cells assigned to singlets, doublets, and negatives
as a function of the positive.quantile provided to Seurat::HTODemux
}
