% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prep_intrahash_doub}
\alias{prep_intrahash_doub}
\title{Estimate intra-hashtag doublets by finding cells whose nearest neighbours are doublets}
\usage{
prep_intrahash_doub(seurat_obj, assay = "RNA", npcs = 20, randomSeed = 12345)
}
\arguments{
\item{seurat_obj}{a Seurat object}

\item{assay}{Seurat object assay to use}

\item{npcs}{number of principal components to use}

\item{randomSeed}{random seed for random number generator}
}
\value{
Seurat object with a predicted_dub_intra logical column added to metadata
}
\description{
Estimate intra-hashtag doublets by finding cells whose nearest neighbours are doublets
}
