% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{prep_qc_rna}
\alias{prep_qc_rna}
\title{Perform QC on RNA, classifying cells as keepers or outliers to be filtered}
\usage{
prep_qc_rna(seurat_obj, assay = "RNA")
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{assay}{assay to use}
}
\value{
Seurat object with logical qc_discard column added to metadata
}
\description{
Perform QC on RNA, classifying cells as keepers or outliers to be filtered
}
