% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_detachAllPkgs}
\alias{utils_detachAllPkgs}
\title{Detach all packages from current session}
\usage{
utils_detachAllPkgs()
}
\value{
None
}
\description{
Detach all packages from current session
}
\examples{
utils_detachAllPkgs()
}
\references{
https://stackoverflow.com/questions/7505547/detach-all-packages-while-working-in-r
}
