% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_detectCoresRAM}
\alias{utils_detectCoresRAM}
\title{Detect cores with RAM constraints}
\usage{
utils_detectCoresRAM(Gb_max = 250, additional_Gb = 1)
}
\arguments{
\item{Gb_max}{ceiling on session memory usage in Gb, assuming that each worker duplicates the session memory}

\item{additional_Gb}{max additional memory requirement for new (temporary) objects created within a parallel session}
}
\value{
max number of cores (integer)
}
\description{
Given a ceiling on RAM to use and predicted extra RAM needed in
. each worker in addition to current usage, return safe number of cores
}
\examples{
n_cores_use = utils_detectCoresRAM(Gb_max=250,additional_Gb=1)
}
