% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_dlFile}
\alias{utils_dlFile}
\title{Check if a file matches given hash, else download it and check again}
\usage{
utils_dlFile(url, destfile, correct_checksum)
}
\arguments{
\item{url}{passed to download.file()}

\item{destfile}{passed to download.file().}

\item{correct_checksum}{file md5 checksum}
}
\value{

}
\description{
Check if a file matches given hash, else download it and check again
}
\examples{
utils_dlFile(url=https:/.mysite.com/files/test.csv, destfile="./data", correct_checksum="2c3294272a968ffc91c0c2af2c296988")
}
