% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_printObjectSizes}
\alias{utils_printObjectSizes}
\title{Show the memory size of the top n objects}
\usage{
utils_printObjectSizes(n)
}
\arguments{
\item{n}{how many objects to list}
}
\value{

}
\description{
Show the memory size of the top n objects
}
\examples{
utils_printObjectSizes(10)
}
