% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_saveMeta}
\alias{utils_saveMeta}
\title{Save file with metadata}
\usage{
utils_saveMeta(savefnc = NULL, doPrint = F, path_log = NULL, ...)
}
\arguments{
\item{savefnc}{a function to write some file to disk. If given as object is converted to character, default NULL}

\item{doPrint}{print output to screen? useful if directin Rscript stdout to a log file (&>), default F}

\item{path_log}{specify log file path; defaults to creating a file in current working dir, default NULL}

\item{\dots}{named arguments to savefnc}
}
\value{
None
}
\description{
saves the following metadata
. filename (if savefnc is not NULL) or else
. current date and time
. sessionInfo() #sessioninfo::session_info()
. github log, if current or parent dir is a github dir
}
\examples{
utils_saveMeta(savefnc=ggplot2::ggsave, doPrint=F, path_log = "~/log.txt", width=10)
TODO add pander pandoc markdown output option?
TODO add MD5SUM save
TODO add overwrite argument
}
