% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils_verboseFnc}
\alias{utils_verboseFnc}
\title{Wrapper to make a function more verbose}
\usage{
utils_verboseFnc(fnc, ...)
}
\arguments{
\item{fnc}{function to execute}

\item{\dots}{named arguments to function}
}
\value{
value returned by fnc
}
\description{
Wrapper to make a function more verbose
}
\examples{
utils_verboseFnc(rowSums, x=Matrix::Matrix(data=rnorm(9), nrow = r), na.rm = TRUE)
}
